Simplify and defer service, allowing it to be injected by Laravel
