+++ src/ActiveCampaignServiceProvider.php
@@ -2,14 +2,15 @@
 
 namespace Label84\ActiveCampaign;
 
+use Illuminate\Contracts\Support\DeferrableProvider;
 use Illuminate\Support\ServiceProvider;
 
-class ActiveCampaignServiceProvider extends ServiceProvider
+class ActiveCampaignServiceProvider extends ServiceProvider implements DeferrableProvider
 {
     public function register(): void
     {
         $this->app->singleton(ActiveCampaign::class, function () {
-            return new ActiveCampaignService(
+            return new ActiveCampaign(
                 config('active-campaign.base_url'),
                 config('active-campaign.api_key'),
                 config('active-campaign.timeout'),
@@ -25,21 +26,18 @@ class ActiveCampaignServiceProvider extends ServiceProvider
         $this->publishes([$configPath => $this->getConfigPath()], 'config');
     }
 
-    /**
-     * Get the config path
-     */
     protected function getConfigPath(): string
     {
         return config_path('active-campaign.php');
     }
 
     /**
-     * Publish the config file
+     * Defer provision of ActiveCampaign service until needed
      *
-     * @param  string  $configPath
+     * @return array<int, string>
      */
-    protected function publishConfig($configPath): void
+    public function provides(): array
     {
-        $this->publishes([$configPath => config_path('active-campaign.php')], 'config');
+        return [ActiveCampaign::class];
     }
 }
